moved base interface logic to the service layer
